Browse styles and header padding
Added some simple styles for Browse (handlebars templates need to be
recompiled to see the Browse styles fully). Added some javascript to
add space for phone status bar if the device was using ios7 or greater.
This  javascript will probably have to be altered to work with the
refactored branch. Made the header and footer opaque.